refactor: move shutdown_process & co to vllm.py
The backends module, along with the llama_cpp and vllm modules,
have a circular dependencies, which is considered an anti-pattern.

Move shutdown_process to eliminate the circular dependency
vllm.py -> backends.
